Add input/response blocks (#4)
This PR adds input and response block support, modeled on Open edX and LON-CAPA. 

This is not end-to-end finished, as there is, for example:

* No user-facing feedback on correctness.
* We don't have the concept of a problem with multiple responses. <Problem> needs to become an action too.
* Data formats are far from finalized
* Ton of TODOs left
* We might want to make a lot of renames (e.g. ResponseType -> Grader, and a translation later OLX 1 -> OLX 2)
* Async graders / split between submission and grading events

However, this is a major step forward, as we can traverse the OLX DOM, infer input relations to responses, run a grader, etc. from where we can keep going incrementally.